feat: Cleaning up the code
